IDVA6-2271 Another few tests refactored
